% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{addCor}
\alias{addCor}
\title{Add correlation of ChIP-seq coverage to motif pairs.}
\usage{
addCor(gi, bwFile, name = "chip", window = 1000, binSize = 1)
}
\arguments{
\item{gi}{\code{\link[InteractionSet]{GInteractions}} object.}

\item{bwFile}{File path or connection to BigWig file with ChIP-seq signals.}

\item{name}{Character indicating the sample name.}

\item{window}{the window size around the center of ranges in \code{gr}.}

\item{binSize}{size of bins to which the coverage values are combined.}
}
\description{
This function first adds ChIP-seq signals along all regions of motif location
using the function \code{\link{addCovToGR}}. Than it calculates the
correlation of coverage for each input pair using the function
\code{\link{addCovCor}}. The Pearson correlation coeffiicent is added as new
metadata column to the input interactions.
}
