% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChIP-seq.R
\name{addCovToGR}
\alias{addCovToGR}
\title{Add coverage to regions in \code{\link[GenomicRanges]{GRanges}} object.}
\usage{
addCovToGR(gr, bwFile, window = 1000, binSize = 1, colname = "cov")
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges}} object with genomic regions}

\item{bwFile}{File path or connection to BigWig file with coverage to parse
from.}

\item{window}{the window size around the center of ranges in \code{gr}.}

\item{binSize}{size of bins to which the coverage values are combined.}

\item{colname}{name of the new column that is created in \code{gr}.}
}
\value{
\code{\link[GenomicRanges]{GRanges}} as input but with an additional
  meta column containing the coverage values for each region.
}
\description{
This function adds a vector of coverage (or any other signal in the input
bigWig file) to each range in a \code{\link[GenomicRanges]{GRanges}} object.
The coverage is reported for a fixed-sized window around the region center.
For regions with negative strand, the coverage vector is reversed.
}
