% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{addInteractionSupport}
\alias{addInteractionSupport}
\title{Add column to \code{\link{GInteractions}} with overlap support.}
\usage{
addInteractionSupport(gi, subject, colname = "loop", ...)
}
\arguments{
\item{gi}{\code{\link{GInteractions}} object}

\item{subject}{another \code{\link{GInteractions}} object}

\item{colname}{name of the new annotation column in \code{gi}.}

\item{...}{additional arguments passed to \code{\link[IRanges]{overlapsAny}}.}
}
\value{
\code{\link{InteractionSet}} \code{gi} as input but with additional
  annotation column \code{colname} indicating whether each interaction
  is supported by \code{subject} or not.
}
\description{
See overlap methods in \code{\link{InteractionSet}} package for more details
on the overlap calculations: \code{?InteractionSet::overlapsAny}
}
