% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{applyToCloseGI}
\alias{applyToCloseGI}
\title{Apply a function to pairs of close genomic regions.}
\usage{
applyToCloseGI(gi, datcol, fun = cor, colname = "value", maxDist = NULL)
}
\arguments{
\item{gi}{A sorted \code{\link[InteractionSet]{GInteractions}} object.}

\item{datcol}{a string matching an annotation column in \code{regions(gi)}.
This collumn is assumed to hold the same number of values for each
interaction \code{NumericList}.}

\item{fun}{A function that takes two numeric vectors as imput to compute a
summary statsitic. Default is \code{\link{cor()}}.}

\item{colname}{A string that is used as columnname for the new column in
\code{gi}.}

\item{maxDist}{maximal distance of pairs in bp as numeric. If maxDist=NULL,
the maximal distance is computed from input interactions gi by
\code{max(pairdist(gi))}.}
}
\value{
A \code{\link[InteractionSet]{GInteractions}} similar to \code{gi}
  just wiht an additinoal column added.
}
\description{
This function adds a vector with the resulting functin call for each input
interaction. Only works for input interaction within the given \code{maxDist}
distance
}
