% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromloop.R
\docType{package}
\name{chromloop}
\alias{chromloop}
\alias{chromloop-package}
\title{Prediction of Chromatin Looping Interaction from ChIP-seq and Sequnece
Motifs.}
\description{
Chromatin looping interactions are an importent feature of eukaryotic genomes
and can bring regulatory sequences such as enhancers in close physical
proximity of regulated target genes.
Here we provide a tool that uses protein
binding signals from ChIP-seq and sequence motif information to predict
chromatin looping events. It can be used to associate binding sites of
transcription factors to regulated genes.
}
