% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cutoffByTF}
\alias{cutoffByTF}
\title{Optimal cutoff values for logistic regression models.}
\format{An object of class \code{tbl_df} with 121 rows and 3 columns:

  \describe{

  \item{TF}{Transcription factor name}

  \item{max_cutoff}{The optimal cutoff on the logistic regression response
  value}

  \item{max_f1}{The optimal f1-score associated to the \code{max_cutoff}
  value}

  }}
\usage{
cutoffByTF
}
\description{
This dataset contains optimal cutoff scores for the response value of logistic
regression models. The cutoff is based on optimal F1-scores. A separate model
was trained For each of 124 TF ChIP-seq datasets in human GM12878 cells. The
model performance was calculated with Hi-C and ChIA-PET interactions using
10-fold cross-validation.
}
\seealso{
\code{\link{modelBest10Avg}} and \code{\link{TFspecificModels}}
}
\keyword{datasets}
