% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{extendAnchors}
\alias{extendAnchors}
\title{Extend interaction anchors with specific term for inner and outer site}
\usage{
extendAnchors(gi, inner, outer)
}
\arguments{
\item{gi}{\code{\link{GInteractions}} or \code{\link{InteractionSet}} object}

\item{inner, outer}{A scalar, non-negative, of the extension sizes for anchor
ends outside and inside interaction loops.}
}
\value{
gi \code{\link{GInteractions}} object with extened regions. The
  \code{\link{GInteractions}} object has two metadata columns, \code{dist},
  and \code{dist_log10} with the genomic distance between the midpointsof
  anchors and distance in log10 space, respectively.
}
\description{
Extend interaction anchors with specific term for inner and outer site
}
