% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{getCisPairs}
\alias{getCisPairs}
\title{Build a \code{\link{GInteractions}} object with all pairs of input
\code{\link{GRanges}} within a given distance.}
\usage{
getCisPairs(inGR, maxDist = 10^6)
}
\arguments{
\item{inGR}{\code{\link{GRanges}} object of genomic regions. The ranges shuld
be sorted according to chr, strand, and start position. Use
\code{\link[GenomicRanges]{sort}} to sort it.}

\item{maxDist}{maximal distance in base-pairs between pairs of ranges as
single  numeric value.}
}
\value{
A \code{\link[InteractionSet]{GInteractions}} object with all pairs
  within the given distance.
}
\description{
Distance is calculated from the center of input regions.
}
