% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_associations.R
\name{linkRegionsInLoops}
\alias{linkRegionsInLoops}
\title{Associate two sets of regions together when they are within a loop.}
\usage{
linkRegionsInLoops(gr1, gr2, gi, ...)
}
\arguments{
\item{gr1}{A GRanges object.}

\item{gr2}{A GRanges object.}

\item{gi}{A \code{\link[InteractionSet]{GInteractions}} or
\code{\link[InteractionSet]{InteractionSet}} object.}

\item{...}{Other arguments passed to \code{\link[GenomicRanges]{findOverlaps}}
and \code{\link[InteractionSet]{linkOverlaps}}}
}
\value{
A dataframe of integer indices indicating which elements of
  \code{gr1} link which elements of \code{gr1}.
}
\description{
Associate two sets of regions together when they are within a loop.
}
