% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modelBest10Avg}
\alias{modelBest10Avg}
\title{Default parameters for logistic regression model.}
\format{An object of class \code{data.frame} with 7 rows and 2 columns:

  \describe{

  \item{term}{Term name}

  \item{estimate}{Parameter estimate}

  }}
\usage{
modelBest10Avg
}
\description{
This dataset contain term names and estimates for logistic regression model
to predict chromatin looping interactions. The estimate represent an average
of the 10 best performing models out of 121 transcription factor ChIP-seq
data sets from ENCODE.
}
\details{
Each of 121 transcritpion factor ChIP-seq data sets from ENCODE in GM12878
cells were used to train a logistic regression model. All CTCF moifs in
\code{\link{motif.hg19.CTCF}} wihtin a distance of 1 Mb were used as
candidates. A given pair was labeld as true loop interactions, if it has
interaction support based on Hi-C lops in GM12878 from Rao et al. 2014 or
ChIA-PET loops from Tang et al. 2015 in GM12878 cells. The 10 best perfroming
models were selcted based on the average area under the
precision-recall-cureve in 10-fold cross-validation.
}
\keyword{datasets}
