% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif.hg19.CTCF}
\alias{motif.hg19.CTCF}
\title{CTCF motif locations in human genome hg19.}
\format{\link[GenomicRanges]{\code{GRanges}} object with 132739 ranges on
  positive and negative strnad indicated and 5 meta columns:
\describe{
  \item{sequence}{genomic DNA sequnece}
  \item{weight}{weight of the diamond, in carats}
}}
\source{
\url{http://jaspar.genereg.net/html/DOWNLOAD/JASPAR_CORE/pfm/individual/MA0139.1.pfm} and \url{http://www.rsat.eu/}
}
\usage{
motif.hg19.CTCF
}
\description{
A dataset containing the motif hits of the CTCF recogintion motif from JASPAR
database (MA0139.1) in human genome assembly hg19. Only motifs with a p-value
threshold of $ \leq $ 0.0001 are reported.
}
\details{
The dataset was created by using the tool RSTA matrix-scan
(\url{http://www.rsat.eu/}) to get all motif hits in the hg19 genome sequence.
The PFM motif model for CTCF was downloaded from JASPAR database using this
url:
\url{http://jaspar.genereg.net/html/DOWNLOAD/JASPAR_CORE/pfm/individual/MA0139.1.pfm}
The following parameters where used.

\describe{
  \item{-quci}{}
  \item{origin start}{}
  \item{-bginput}{}
  \item{-markov 1}{}
  \item{-2str}{}
  \item{-uth pval 0.0001}{}
  \item{-return pvalue}{}
}
}
\keyword{datasets}
