% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{noZeroVar}
\alias{noZeroVar}
\title{returns indecies of columns with non-zero variance}
\usage{
noZeroVar(dat)
}
\arguments{
\item{dat}{data.frame or matirx}
}
\value{
column indecies of columns with non-zero variance
}
\description{
returns indecies of columns with non-zero variance
}
