% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChIP-seq.R
\name{parseBAMToRle}
\alias{parseBAMToRle}
\title{Parse BAM file as RleList coverage object.}
\usage{
parseBAMToRle(inFile, seqInfo, selectionGR = NULL)
}
\arguments{
\item{inFile}{a BAM file}

\item{seqInfo}{a \code{\link[GenomeInfoDb]{seqinfo}} object.}

\item{selectionGR}{\code{\link[GenomicRanges]{GRanges}} object with regions for
which the coverage is selected.}
}
\value{
\code{\link[S4Vectors]{Rle-class}} coverage object. See
 \code{\link[IRanges]{coverage}}.
}
\description{
See http://bioconductor.org/packages/release/bioc/html/Rsamtools.html
}
