% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{pred_logit}
\alias{pred_logit}
\title{Predict interaction probability using logistic regresstion model.}
\usage{
pred_logit(data, formula, betas)
}
\arguments{
\item{data}{A data.frame with predictor variables}

\item{formula}{A modelling formula. All predictor variables should be availabel in the data frame.}

\item{betas}{A vector with parameter estimates for predictor variables}
}
\value{
A numeric vector with interaction probabilites for each ovservation
  in \code{df}.
}
\description{
Predict interaction probability using logistic regresstion model.
}
