% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{prepareCandidates}
\alias{prepareCandidates}
\title{Prepares motif pairs and add genomic features.}
\usage{
prepareCandidates(motifs, bwFile, maxDist = 10^6, window = 1000,
  binSize = 10, motifScoreCol = "sig")
}
\arguments{
\item{moitfs}{GRanges object with motif locations}
}
\description{
Prepares motif pairs and add genomic features.
}
