% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChIP-seq.R
\name{slideMean}
\alias{slideMean}
\title{Sliding mean over x of intervals of size k}
\usage{
slideMean(x, k)
}
\arguments{
\item{x}{numeric vector}

\item{k}{interval size}
}
\value{
numeric vector of length \code{length(x) / k}.
}
\description{
Source:
\url{http://stats.stackexchange.com/questions/3051/mean-of-a-sliding-window-in-r}
}
