% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{addCor}
\alias{addCor}
\title{Prepares motif pairs and add genomic features.}
\usage{
addCor(motifs, bwFile, maxDist = 10^6, window = 1000, binSize = 10,
  motifScoreCol = "sig")
}
\arguments{
\item{bwFile}{File path or connection to BigWig file with coverage to
parrse from.}

\item{gi}{GInteractions object}
}
\description{
Prepares motif pairs and add genomic features.
}
