% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{addCovCor}
\alias{addCovCor}
\title{Add correlation of anchor signals to pairs of close genomic regions.}
\usage{
addCovCor(gi, datcol, colname = "cor", maxDist = NULL)
}
\arguments{
\item{gi}{A sorted \code{\link[InteractionSet]{GInteractions}} object.}

\item{datcol}{a string matching an annotation column in \code{regions(gi)}.
This column is assumed to hold the same number of values for each
interaction as a \code{NumericList}.}

\item{colname}{A string that is used as columnname for the new column in
\code{gi}.}

\item{maxDist}{maximal distance of pairs in bp as numeric. If maxDist=NULL,
the maximal distance is computed from input interactions gi by
\code{max(pairdist(gi))}.}
}
\value{
A \code{\link[InteractionSet]{GInteractions}} similar to \code{gi}
  just with an additional column added.
}
\description{
This function adds a vector with correlation values for each input
interaction. Only works for input interaction within the given \code{maxDist}
distance.
}
