% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChIP-seq.R
\name{addCovToGR}
\alias{addCovToGR}
\title{Add coverage to regions in \code{\link[GenomicRanges]{GRanges}} object.}
\usage{
addCovToGR(gr, cov, window = 1000, bin_size = 10, colname = "cov")
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges}} object with genomic regions}

\item{cov}{\code{\link{RleList}} object with coverage for each chromosome. Such
an object is returned from \code{\link{parseBigWigToRle}} function.}

\item{window}{the window size arund the center of ranges in \code{gr}.}

\item{bin_size}{size of bins to which the coverage values are combined.}
}
\value{
\code{\link[GenomicRanges]{GRanges}} as input but with an additional meta column containing the coverage values for each region.
}
\description{
This function adds a coverage vector to each range in a genomic ranges
object. The coverage is reported for a fixed-sized window around the region
center and is reversed in case of negative strand of the region.
}
