% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChIP-seq.R
\name{addCovToGR}
\alias{addCovToGR}
\title{Add coverage to regions in \code{\link[GenomicRanges]{GRanges}} object.}
\usage{
addCovToGR(gr, bwFile, window = 1000, binSize = 1, colname = "cov")
}
\arguments{
\item{gr}{\code{\link[GenomicRanges]{GRanges}} object with genomic regions}

\item{bwFile}{File path or connection to BigWig file with coverage to parse
from.}

\item{window}{Numeric scalar for window size around the center of ranges in
\code{gr}.}

\item{binSize}{Integer scalar as size of bins to which the coverage values are
combined.}

\item{colname}{Character as name of the new column that is created in
\code{gr}.}
}
\value{
\code{\link[GenomicRanges]{GRanges}} as input but with an additional
 meta column containing the coverage values for each region as
 \code{\link[IRanges]{NumericList}}.
}
\description{
This function adds a \code{\link[IRanges]{NumericList}}  of coverage (or any
other signal in the input bigWig file) to each range in a
\code{\link[GenomicRanges]{GRanges}} object. The coverage is reported for a
fixed-sized window around the region center. For regions with negative strand,
the coverage vector is reversed. The coverage signal is added as new metadata
colum holding a \code{\link[IRanges]{NumericList}} object.
This function does not work on Windows. Only NAs are added on Windows.
}
\examples{

# use example bigWig file of ChIP-seq signals on human chromosome 22
exampleBigWig <- system.file("extdata",
"GM12878_Stat1.chr22_1-18000000.bigWig", package = "chromloop")

# use example CTCF moitf location on human chromosome 22
motifGR <- chromloop::motif.hg19.CTCF.chr22

# add ChIP-seq signals to motif regions
motifGR <- addCovToGR(motifGR, exampleBigWig)

# add ChIP-seq signals as column named "Stat1"
motifGR <- addCovToGR(motifGR, exampleBigWig, colname = "Stat1")

# add ChIP-seq signals in windows of 500bp around motif centers
motifGR <- addCovToGR(motifGR, exampleBigWig, window = 500)

# add ChIP-seq signals in bins of 10 bp
motifGR <- addCovToGR(motifGR, exampleBigWig, binSize = 10)

}
