% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{addInteractionSupport}
\alias{addInteractionSupport}
\title{Add column to \code{\link{GInteractions}} with overlap support.}
\usage{
addInteractionSupport(gi, subject, colname = "loop", ...)
}
\arguments{
\item{gi}{\code{\link{GInteractions}} object}

\item{subject}{another \code{\link{GInteractions}} object}

\item{colname}{name of the new annotation columm in \code{gi}.}

\item{...}{addtional arguments passed to \code{\link{IRanges::overlapsAny}}.}
}
\value{
\code{\link{InteractionSet}} \code{gi} as input but with additonal
  annotation column \code{colname} indicationg whether each interaction
  is supported by \code{subject} or not.
}
\description{
See overlap methods in \code{\link{InteractionSet}} package for more details
on the oberlap calculations: \code{?InteractionSet::overlapsAny}
}
