% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{addMotifScore}
\alias{addMotifScore}
\title{Add motif score of anchors.}
\usage{
addMotifScore(gi, colname = "score")
}
\arguments{
\item{gi}{\code{\link{GInteractions}}}

\item{colname}{name of the new colum that is created in \code{gi}.}
}
\value{
The same \code{\link{GInteractions}} as \code{gi} but with three
  additonal annotation columns.
}
\description{
If each anchor region (motif) has a score as annotation column, this function
adds two new columns named "score_1" and "score_2" with the scores of the
first and the second anchor region, respectively. Additonally a column named
"score_min" is added with holds for each interaction the mimium of "score_1"
and "score_2".
}
