% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{addStrandCombination}
\alias{addStrandCombination}
\title{Add combination of anchor strand orientation.}
\usage{
addStrandCombination(gi, colname = "strandOrientation")
}
\arguments{
\item{gi}{\code{\link{GInteractions}}}

\item{colname}{name of the new column that is created in \code{gi}.}
}
\value{
The same \code{\link{GInteractions}} as \code{gi} but with an
  additional column indicating the four possible combinations of strands
  "forward", "reverse", "convergent", or "divergent".
}
\description{
Each anchor region has a strand that is \code{'+'} or \code{'-'}. Therefore,
the each interaction between two regions has one of the following strand
combinations: "forward", "reverse", "convergent", or "divergent". Unstranded
ranges, indicated by \code{*}, are treated as positive strand.
}
