% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cutoffBest10}
\alias{cutoffBest10}
\title{Default optimal cutoff value of logistic regression.}
\format{An object of class \code{numeric} of length 1.}
\usage{
cutoffBest10
}
\description{
This value is the average optimal cutoff value on the 10 best performing TF
ChIP-seq data sets. It is used as default cutoff value on the logistic
regression response score in \code{\link{predLoops}} function See
\code{?'cutoffByTF'} for more details.
}
\seealso{
\code{\link{cutoffByTF}}, \code{\link{modelBest10Avg}}
}
\keyword{datasets}
