% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChIP-seq.R
\name{getOutOfBound}
\alias{getOutOfBound}
\title{Get out of chromosomal bound ranges.}
\usage{
getOutOfBound(gr)
}
\arguments{
\item{gr}{A \code{GRanges} object.}
}
\value{
A \code{data.frame} with rows for each range in \code{gr} that
  extends out of chromosomes. The first column holds the index of the range in
  \code{gr}, the second the size of the overlap to the left of the chromosome
  and the third the size of the overlap to the right of the chromosome.
}
\description{
Get out of chromosomal bound ranges.
}
\keyword{internal}
