% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{interactionRange}
\alias{interactionRange}
\title{Get genomic ranges spanned by (intracrhomosomal) interactions.}
\usage{
interactionRange(gi)
}
\arguments{
\item{gi}{\code{\link[InteractionSet]{GInteractions}} or
\code{\link[InteractionSet]{InteractionSet}} object}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object with ranges spanning
  the interactions in \code{gi}.
}
\description{
Note, this method assumes are intra-chromosomal (e.g. only interactions
between regions from the same chromosome).
}
