% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_associations.R
\name{linkRegions}
\alias{linkRegions}
\title{Associate two sets of regions together by interactinos.}
\usage{
linkRegions(gr1, gr2, gi, ignore.strand = TRUE, ...)
}
\arguments{
\item{gr1}{A GRanges object.}

\item{gr2}{A GRanges object.}

\item{gi}{A \code{\link[InteractionSet]{GInteractions}} or
\code{\link[InteractionSet]{InteractionSet}} object.}

\item{ignore.strand}{Logical. See \code{\link[InteractionSet]{linkOverlaps}}.}

\item{...}{Other arguments passed to \code{\link[GenomicRanges]{findOverlaps}}
and \code{\link[InteractionSet]{linkOverlaps}}}
}
\value{
A dataframe of integer indices indicating which elements of
  \code{gr1} link which elements of \code{gr1}.
}
\description{
Associate two sets of regions together by interactinos.
}
