% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChIP-seq.R
\name{parseBigWigCov}
\alias{parseBigWigCov}
\title{Parse coverage for specific regions from bigWig file.}
\usage{
parseBigWigCov(inFile, as = "RleList", ...)
}
\arguments{
\item{inFile}{The connection from which data is loaded. If this is a
character vector, it is assumed to be a filename and a corresponding file
connection is created}

\item{...}{Other parameters to pass to \code{\link[rtracklayer]{import.bw}}.}
}
\description{
Parse coverage for specific regions from bigWig file.
}
