% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseChIP-seq.R
\name{parseBigWigToRle}
\alias{parseBigWigToRle}
\title{The bigWig files contains read counts (or other dense, continouse data)
along the genome.
The bigWig format is described here:
https://genome.ucsc.edu/goldenpath/help/bigWig.html}
\usage{
parseBigWigToRle(inFile, seqInfo, format = "bigWig")
}
\arguments{
\item{inFile}{Input file path or connection. See \code{con} paramter in
\code{\link[rtracklayer]{import}} function.}

\item{seqInfo}{A \link{\code{seqinfo}} object defining the reference genome.}
}
\value{
An \code{\link{RleList}} object with density values for each
position in the genome.
}
\description{
The bigWig files contains read counts (or other dense, continouse data)
along the genome.
The bigWig format is described here:
https://genome.ucsc.edu/goldenpath/help/bigWig.html
}
