% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseInteractions.R
\name{parseLoopsRao}
\alias{parseLoopsRao}
\title{Parse chromatin loops from Rao et al. 2014 as strict
\code{\link[InteractionSet]{GInteractions}}.}
\usage{
parseLoopsRao(inFile, ...)
}
\arguments{
\item{inFile}{input file with loops}

\item{...}{additional arguments, that will be passed to
\code{\link[GenomicRanges]{GRanges}} functions.}
}
\value{
\code{\link{GInteractions}} with loops from input file.
}
\description{
Parse chromatin loops from Rao et al. 2014 as strict
\code{\link[InteractionSet]{GInteractions}}.
}
