% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseInteractions.R
\name{parseLoopsRao}
\alias{parseLoopsRao}
\title{Parse loops from Rao et al 2014 as two \code{\link{InteractionSet}} object.}
\usage{
parseLoopsRao(inFile, ...)
}
\arguments{
\item{...}{additional arguments, that will be passed to
\code{\link{GRanges()}} functions.}

\item{infile}{input file with loops}
}
\value{
An code{\link{InteractionSet}} with loops from input file.
}
\description{
Parse loops from Rao et al 2014 as two \code{\link{InteractionSet}} object.
}
