% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{pred_logit}
\alias{pred_logit}
\title{Predict interaction probability using logistic regresstion model.}
\usage{
pred_logit(data, formula, betas)
}
\arguments{
\item{data}{A data.frame like object with predictor variables.}

\item{formula}{A \code{\link[stats]{formula}}. All predictor variables should
be available in the data frame.}

\item{betas}{A vector with parameter estimates for predictor variables. They
should be in the same order as variables in \code{formula}.}
}
\value{
A numeric vector with interaction probabilities for each ovservation
  in \code{df}. NAs are produced for NAs in \code{df}.
}
\description{
Predict interaction probability using logistic regresstion model.
}
