% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predLoops}
\alias{predLoops}
\title{Predict looping interaction probability.}
\usage{
predLoops(gi, formula = NULL, betas = NULL, colname = "pred")
}
\arguments{
\item{gi}{A \code{\link[InteractionSet]{GInteractions}} object with coverage
corealtion and genomic features in metadata columns. See
\link{prepareCandidates} and \link{addCor} to build it.}

\item{formula}{A \code{\link[stats]{formula}}. All predictor variables should
be availabel in the in metadat colums of \code{gi}. If NULL, the following
default formular is used: \code{~ dist + strandOrientation + score_min +
chip}.}

\item{betas}{A vector with parameter estimates for predictor variables. They
should be in the same order as variables in \code{formula}.}

\item{colname}{A \code{character} as column name of new metadata colum in
\code{gi} for predictions.}
}
\value{
A \code{\link[InteractionSet]{GInteractions}} as \code{gi} with an
  additional metadata colum holdin the predicted looping probability.
}
\description{
Predict looping interaction probability.
}
\seealso{
\code{\link{prepareCandidates}}, \code{\link{addCor}},
  \code{\link{pred_logit}}
}
