% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{prepareCandidates}
\alias{prepareCandidates}
\title{Prepares motif pairs and add genomic features.}
\usage{
prepareCandidates(motifs, maxDist = 1e+06, scoreColname = "score")
}
\arguments{
\item{motifs}{\code{\link[GenomicRanges]{GRanges}} object with motif
locations.}

\item{maxDist}{maximal distance in base-pairs between pairs of ranges as
single  numeric value.}

\item{scoreColname}{Character as name the metadata column in with motif
score.}
}
\value{
An \code{\link[InteractionSet]{GInteractions}} object with motif
  pairs and annotations of distance, strand orientation, and motif scores.
}
\description{
Prepares motif pairs and add genomic features.
}
\examples{

# build example GRanges as anchors
anchorGR <- GRanges(
 rep("chr1", 4),
 IRanges(
   c(1, 5, 20, 14),
   c(4, 8, 23, 17)
 ),
 strand = c("+", "+", "+", "-"),
 score = c(5, 4, 6, 7)
)



# prepare candidates
gi <- prepareCandidates(anchorGR)


# prepare candidates using a mimial distance of 10 bp
gi <- prepareCandidates(anchorGR, maxDist = 10)

# prepare candidates using an alternative score value in anchors
anchorGR$myScore <- rnorm(length(anchorGR))
gi <- prepareCandidates(anchorGR, scoreColname = "myScore")

}
