% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{prepareCandidates}
\alias{prepareCandidates}
\title{Prepares motif pairs and add genomic features.}
\usage{
prepareCandidates(motifs, maxDist = 1e+07, scoreColname = "score")
}
\arguments{
\item{motifs}{\code{\link[GenomicRanges]{GRanges}} object with motif
locations.}

\item{maxDist}{maximal distance in base-pairs between pairs of ranges as
single  numeric value.}

\item{scoreColname}{Character as name the metadata column in with motif
score.}
}
\value{
An \code{\link[InteractionSet]{GInteractions}} object with motif
  pairs and annotations of distance, strand orientation, and motif scores.
}
\description{
Prepares motif pairs and add genomic features.
}
